crawling nicely

+++ src/easy_robot_control/launch/launch_setting.py
@@ -13,7 +13,7 @@ from launch.substitutions import (
 std_movement_time = 2  # seconds
 movement_update_rate = 30.0  # Hz
 number_of_legs = 1
-wheel_size = float(210)
+wheel_size = float(230)
 
 ROS2_PACKAGE_WITH_URDF = "rviz_basic"
 # ROBOT_NAME_DEFAULT = "moonbot_7"
